`expand_matrix` moved from `operation.py` --> `math/matrix_manipulation.py` (#3008)
* expand_matrix moved:  operation --> math/matrix_manipulation

* lint + changelog

* moved to breaking changes

* skip docs build for imports

* typo
