+++ doc/releases/changelog-dev.md
@@ -213,6 +213,10 @@
   the new `circuit` keyword argument.
   [(#2820)](https://github.com/PennyLaneAI/pennylane/pull/2820)
 
+* The `expand_matrix()` has been moved from `~/operation.py` to 
+  `~/math/matrix_manipulation.py`
+  [(#3008)](https://github.com/PennyLaneAI/pennylane/pull/3008)
+
 <h3>Deprecations</h3>
 
 * The `supports_reversible_diff` device capability is unused and has been removed.
