+++ pennylane/ops/op_math/controlled_class.py
@@ -289,11 +289,11 @@ class Controlled(SymbolicOp):
         canonical_matrix = qmlmath.block_diag([left_pad, base_matrix, right_pad])
 
         if wire_order is None or self.wires == Wires(wire_order):
-            return operation.expand_matrix(
+            return qml.math.expand_matrix(
                 canonical_matrix, wires=self.active_wires, wire_order=self.wires
             )
 
-        return operation.expand_matrix(
+        return qml.math.expand_matrix(
             canonical_matrix, wires=self.active_wires, wire_order=wire_order
         )
