+++ tests/devices/test_default_qubit_tf.py
@@ -261,7 +261,7 @@ class TestTFMatrix:
             expected_mat = qml.math.kron(I, qml.math.kron(reg_mat, I))
 
         tf_mat = op(tf.Variable(param), wires=wires).matrix()
-        obtained_mat = qml.operation.expand_matrix(tf_mat, wires, list(range(4)))
+        obtained_mat = qml.math.expand_matrix(tf_mat, wires, list(range(4)))
 
         assert qml.math.get_interface(obtained_mat) == "tensorflow"
         assert qml.math.allclose(qml.math.unwrap(obtained_mat), expected_mat)
